// src/repository/enum/activity.member.enum.ts
export enum TypeActivityMemberType {
  ADMIN = 'admin',
  SHARE = 'share',
  NORMAL = 'normal',
}

export enum TypeActivityMemberStatus {
  NORMAL = 'normal',
  DONE = 'done',
  DELETE = 'delete',
  WAIT = 'wait',
}

export enum ActivityMemberSoccerPosition {
  GK = 'GK',
  DF = 'DF',
  MF = 'MF',
  FW = 'FW',
  ST = 'ST',
}

export enum ActivityMemberSoccerHow {
  LEFT = 'left',
  RIGHT = 'right',
  BOTH = 'both',
}

export enum ActivityMemberSoccerStyle {
  ST = 'ST',
  DF = 'DF',
}