// src/repository/enum/activity.repository.enum.ts
export enum TypeActivityStatus {
  NORMAL = 'normal',
  DONE = 'done',
  DELETE = 'delete',
  BLOCK = 'block',
}

export enum ActivitySoccerType {
  FULL = 'full',
  HALF = 'half',
  FOOT = 'foot',
  EDUCATION = 'education',
}

export enum ActivitySoccerFormation {
  FORMATION_3142 = '3142',
  FORMATION_343 = '343',
  FORMATION_352 = '352',
  FORMATION_41212 = '41212',
  FORMATION_4141 = '4141',
  FORMATION_4231 = '4231',
  FORMATION_424 = '424',
  FORMATION_433 = '433',
  FORMATION_4411 = '4411',
  FORMATION_442 = '442',
  FORMATION_532 = '532',
  FORMATION_541 = '541',
}